refactor: renamed compound_icon to sp_compound_icon and moved form utils to components
