+++ lib/src/components/sp_compound_icon.dart
